// app/Global/Types.ts
export type XYObject = { x: number; y: number };
export type SizedObject = { width: number; height: number };
export type Rect = { x: number; y: number; width: number; height: number };
export type Skin = {
    name: string;
    scale: number;
    align: string;
    frames: number;
    framesHold: number;
};